third_party/WebKit/LayoutTests/fast/dom/shadow/event-path.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>
<p id="description"></p>
<div id="sandbox"></div>
<pre id="console"></pre>
<div id='a'>
    <div id='b'>
        <div id='c'>
        </div>
    </div>
</div>
<script>
var b = document.getElementById('b');
b.addEventListener('click', function(event) {
    var path = event.path;
    debug(dumpNodeList(path));
    debug('Makes sure that event.path returns static NodeList.');
    path[1] = '';
    debug(dumpNodeList(event.path));
});
var clickEvent = document.createEvent("MouseEvents");
debug("Makes sure that event.path is empty before dispatching the event.");
debug(dumpNodeList(clickEvent.path));
clickEvent.initMouseEvent('click', true, false, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
b.dispatchEvent(clickEvent);
debug("Makes sure that event.path isn't emptified after dispatching the event.");
debug(dumpNodeList(clickEvent.path));
</script>
</body>
</html>
